refactor(ipc): split register.ts into per-namespace handler modules
Cut main-src/ipc/register.ts from 4236 lines to 1980 (-53%) by extracting
~75 IPC handlers into focused modules and lifting the chat pipeline out
of the dispatcher.

New modules:
- agentRuntime.ts: cross-handler helpers (senderWorkspaceRoot,
  workspaceRootsEqual, parseAppWindowSurface, broadcastPluginsChanged)
- chatRuntime.ts: runChatStream + 6 shared state Maps (abortByThread,
  agentRevertSnapshotsByThread, toolApprovalWaiters, mistakeLimitWaiters,
  preflightAbortByThread, threadTitleGenerationVersion) and helpers
  (recordTurnTokenUsageStats, persistAssistantStreamError,
  queueThreadTitleGeneration, resolveManagedAgentLoopOptions,
  activeUsageStatsDir, appendSystemBlock)
- handlers/{app,workspace,fs,shell,git,browser,mcp,plugins,settings,
  terminalExec}Handlers.ts: one file per IPC namespace

Behaviour is unchanged: handlers were moved verbatim and rewired through
register.ts:registerIpc(). Also clears 67 newly-unused imports left
behind in register.ts.

Fix + regression test for ReferenceError shipped in ref/app:
- register.ts referenced workspaceRootsEqual without importing it after
  the helper moved to agentRuntime.ts. esbuild does not catch free
  identifiers and main-src is outside tsconfig.json `include`, so the
  bug only surfaced when threads:listAgentSidebar fired in production.
- Add main-src/ipc/registerSmoke.test.ts: mocks `electron`, loads every
  register*Handlers() module plus registerIpc(), then invokes every
  captured handler body. Any ReferenceError from a missing import now
  fails CI (verified by temporarily reverting the fix - the test caught
  it). 16 new test cases, 561/562 total pass.
